Introduce new setting to configure when to build graph during segment creation (#2007)
* Fix native engine vector format test (#2103)

Previosuly we were not creating hnsw file on segment flush for faiss engine.
After successfully integrating hnsw file creation, we forgot to update unit test.
Here, we will confirm that required files are being created based on field type.

Signed-off-by: Vijayan Balasubramanian <balasvij@amazon.com>

* Introduce a setting to control whether to build graph

Added new updatable index setting "build_vector_data_structure_threshold", which will be
considered when to build braph or not for native engines.
This is noop for lucene. This depends on use lucene format as prerequisite.
We don't need to add flag since it is only enable if lucene format is
already enabled.

Signed-off-by: Vijayan Balasubramanian <balasvij@amazon.com>